CLEM move transmutations into separate cs files and interface to folder
